Moves RasterSequence to own module.
